+++ backend/app/src/main/java/carrot/candy/app/chat/controller/ChatRoomController.java
@@ -1,7 +1,7 @@
 package carrot.candy.app.chat.controller;
 
 import carrot.candy.app.auth.domain.AuthMember;
-import carrot.candy.app.chat.dto.ChatRoomCreateRequest;
+import carrot.candy.app.chat.dto.request.ChatRoomCreateRequest;
 import carrot.candy.app.chat.dto.response.ChatRoomResponse;
 import carrot.candy.app.chat.service.ChatRoomService;
 import carrot.candy.app.common.annotation.PreAuthorize;
