+++ backend/app/src/main/java/carrot/candy/app/chat/dto/request/ChatRoomCreateRequest.java
@@ -1,4 +1,4 @@
-package carrot.candy.app.chat.dto;
+package carrot.candy.app.chat.dto.request;
 
 public record ChatRoomCreateRequest(
         String name,
