+++ backend/app/src/main/java/carrot/candy/app/chat/service/ChatRoomService.java
@@ -3,7 +3,7 @@ package carrot.candy.app.chat.service;
 import carrot.candy.app.auth.domain.AuthMember;
 import carrot.candy.app.chat.domain.chatroom.ChatRoom;
 import carrot.candy.app.chat.domain.chatroom.ChatRoomRepository;
-import carrot.candy.app.chat.dto.ChatRoomCreateRequest;
+import carrot.candy.app.chat.dto.request.ChatRoomCreateRequest;
 import carrot.candy.app.chat.dto.response.ChatRoomResponse;
 import carrot.candy.app.member.domain.Member;
 import carrot.candy.app.member.domain.MemberRepository;
