+++ backend/app/src/main/java/carrot/candy/app/chat/service/MessageService.java
@@ -4,7 +4,7 @@ import carrot.candy.app.chat.domain.chatroom.ChatRoom;
 import carrot.candy.app.chat.domain.chatroom.ChatRoomRepository;
 import carrot.candy.app.chat.domain.message.Message;
 import carrot.candy.app.chat.domain.message.MessageRepository;
-import carrot.candy.app.chat.dto.MessageSendRequest;
+import carrot.candy.app.chat.dto.request.MessageSendRequest;
 import carrot.candy.app.member.domain.Member;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
